rename: 게임 메인 클래스 이름 변경
NumberBaseballGame -> GameRunner
